Move typed dicts to Announce
