+++ BitTornado/Client/Rerequester.py
@@ -1,66 +1,9 @@
 import threading
 import socket
 import random
-from BitTornado.Network.NetworkAddress import IPv4, IPv6
 from BitTornado.Meta.Info import check_type
-from BitTornado.Meta.TypedCollections import TypedDict, TypedList, QueryDict
 from io import StringIO
 from traceback import print_exc
-
-
-class Peer(TypedDict):
-    iptype = IPv4
-    typemap = {'ip': str, 'port': int, 'peer id': bytes}
-
-    def __init__(self, arg):
-        if isinstance(arg, bytes):
-            nbytes = self.iptype.bits // 8
-            arg = {'ip': self.iptype.from_bytes(arg[:nbytes], 'big'),
-                   'port': int.from_bytes(arg[nbytes:], 'big')}
-        TypedDict.__init__(self, arg)
-
-
-class Peer6(Peer):
-    iptype = IPv6
-
-
-class Response(TypedDict):
-    class Peers(TypedList):
-        valtype = Peer
-
-        def __init__(self, arg):
-            if isinstance(arg, bytes):
-                arg = [arg[i:i+6] for i in range(0, len(arg), 6)]
-            TypedList.__init__(self, arg)
-
-    class Peers6(TypedList):
-        valtype = Peer6
-
-        def __init__(self, arg):
-            assert isinstance(arg, bytes)
-            arg = [arg[i:i+18] for i in range(0, len(arg), 18)]
-            TypedList.__init__(self, arg)
-
-    typemap = {'failure reason': str, 'warning message': str, 'interval': int,
-               'min interval': int, 'tracker id': bytes, 'complete': int,
-               'incomplete': int, 'crypto_flags': bytes, 'peers': Peers,
-               'peers6': Peers6}
-    valmap = {str: str.encode}
-
-
-class RequestURL(QueryDict):
-    typemap = {'info_hash': bytes, 'peer_id': bytes, 'port': int,
-               'supportcrypto': bool, 'requirecrypto': bool, 'cryptoport': int,
-               'seed_id': bytes, 'check_seeded': bool, 'uploaded': int,
-               'downloaded': int, 'left': int, 'numwant': int,
-               'no_peer_id': bool, 'compact': bool, 'last': int,
-               'trackerid': bytes, 'event': str, 'tracker': bool,
-               'key': str}
-
-    def __add__(self, newdict):
-        x = self.__class__(self)
-        x.update(newdict)
-        return x
 
 
 class fakeflag:
